separate out retrofit's Api interfaces specific to the endpoint that it will access.
