+++ app/src/main/java/com/lambdaschool/hackathon_portal/di/app/AppComponent.kt
@@ -10,7 +10,6 @@ import javax.inject.Singleton
 @Singleton
 @Component(
     modules = [
-        AppModule::class,
         Auth0Module::class,
         RepositoryModule::class,
         ViewModelsModule::class
