+++ app/src/main/java/com/lambdaschool/hackathon_portal/retrofit/TeamApiInterface.kt
@@ -1,0 +1,5 @@
+package com.lambdaschool.hackathon_portal.retrofit
+
+interface TeamApiInterface {
+
+}
